Migrated some code out of `dataflow::drop_flag_effects` and into its parent module.
(This code is more general purpose than just supporting drop flag elaboration.)
